pyproject.toml
[project]
name = "aura-hub"
version = "0.1.0"
description = "Starter hub for Aura distributed home voice OS"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "boto3>=1.35.0",
  "langgraph>=0.2.35",
  "spotipy>=2.24.0",
  "supabase>=2.7.0",
  "google-api-python-client>=2.145.0",
  "google-auth>=2.35.0",
  "python-dotenv>=1.0.1",
  "pyyaml>=6.0.2",
  "python-dateutil>=2.9.0",
  "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "ruff>=0.6.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
